# [Constraint.Lagrangian.Correction] LinearSolverCC: use FullVector API if detected