--- a/Sofa/Component/Constraint/Lagrangian/Correction/src/sofa/component/constraint/lagrangian/correction/LinearSolverConstraintCorrection.inl
+++ b/Sofa/Component/Constraint/Lagrangian/Correction/src/sofa/component/constraint/lagrangian/correction/LinearSolverConstraintCorrection.inl
@@ -536,7 +536,7 @@ void LinearSolverConstraintCorrection<DataTypes>::resetForUnbuiltResolution(doub
     systemMatrix_buf   = l_linearSolver.get()->getSystemBaseMatrix();
     systemRHVector_buf = l_linearSolver.get()->getSystemRHBaseVector();
     systemLHVector_buf = l_linearSolver.get()->getSystemLHBaseVector();
-    systemLHVector_buf_fullvector = dynamic_cast<linearalgebra::FullVector<SReal>*>(systemLHVector_buf);
+    systemLHVector_buf_fullvector = dynamic_cast<linearalgebra::FullVector<Real>*>(systemLHVector_buf);
 
     constexpr const auto derivDim = Deriv::total_size;
     const unsigned int systemSize = mstate->getSize() * derivDim;