--- a/Sofa/Component/Constraint/Lagrangian/Correction/src/sofa/component/constraint/lagrangian/correction/LinearSolverConstraintCorrection.inl
+++ b/Sofa/Component/Constraint/Lagrangian/Correction/src/sofa/component/constraint/lagrangian/correction/LinearSolverConstraintCorrection.inl
@@ -536,8 +536,9 @@ void LinearSolverConstraintCorrection<DataTypes>::resetForUnbuiltResolution(doub
     systemMatrix_buf   = l_linearSolver.get()->getSystemBaseMatrix();
     systemRHVector_buf = l_linearSolver.get()->getSystemRHBaseVector();
     systemLHVector_buf = l_linearSolver.get()->getSystemLHBaseVector();
+    systemLHVector_buf_fullvector = dynamic_cast<linearalgebra::FullVector<SReal>*>(systemLHVector_buf);
 
-    const unsigned int derivDim = Deriv::size();
+    constexpr const auto derivDim = Deriv::total_size;
     const unsigned int systemSize = mstate->getSize() * derivDim;
     systemRHVector_buf->resize(systemSize) ;
     systemLHVector_buf->resize(systemSize) ;
@@ -575,16 +576,35 @@ template<class DataTypes>
 void LinearSolverConstraintCorrection<DataTypes>::addConstraintDisplacement(double *d, int begin, int end)
 {
     const MatrixDeriv& constraints = mstate->read(core::ConstMatrixDerivId::constraintJacobian())->getValue();
-    const auto derivDim = Deriv::size();
 
     last_disp = begin;
 
-
-    l_linearSolver.get()->partial_solve(Vec_I_list_dof[last_disp], Vec_I_list_dof[last_force], _new_force);
+    l_linearSolver->partial_solve(Vec_I_list_dof[last_disp], Vec_I_list_dof[last_force], _new_force);
 
     _new_force = false;
 
-    const auto positionIntegrationFactor = l_ODESolver.get()->getPositionIntegrationFactor();
+    constexpr auto addConstraintDisplacement_impl = [](double* d, unsigned int id, auto* systemLHVector_buf, double positionIntegrationFactor, unsigned int dof, const Deriv& val)
+    {
+        constexpr const auto derivDim = Deriv::total_size;
+        Deriv disp(type::NOINIT);
+
+        for (Size j = 0; j < derivDim; j++)
+        {
+            if constexpr (std::is_same_v<decltype(systemLHVector_buf), linearalgebra::FullVector<SReal>*>)
+            {
+                disp[j] = (*systemLHVector_buf)[dof * derivDim + j] * positionIntegrationFactor;
+            }
+            else
+            {
+                disp[j] = (Real)(systemLHVector_buf->element(dof * derivDim + j)) * positionIntegrationFactor;
+            }
+        }
+
+        d[id] += val * disp;
+    };
+
+    const auto positionIntegrationFactor = l_ODESolver->getPositionIntegrationFactor();
+
     // TODO => optimisation => for each bloc store J[bloc,dof]
     for (int i = begin; i <= end; i++)
     {
@@ -596,15 +616,14 @@ void LinearSolverConstraintCorrection<DataTypes>::addConstraintDisplacement(doub
 
             for (MatrixDerivColConstIterator colIt = rowIt.begin(); colIt != rowEnd; ++colIt)
             {
-                const auto dof = colIt.index();
-                Deriv disp(type::NOINIT);
-
-                for(Size j = 0; j < derivDim; j++)
+                if (systemLHVector_buf_fullvector)
                 {
-                    disp[j] = (Real)(systemLHVector_buf->element(dof * derivDim + j) * positionIntegrationFactor);
+                    addConstraintDisplacement_impl(d, i, systemLHVector_buf_fullvector, positionIntegrationFactor, colIt.index(), colIt.val());
                 }
-
-                d[i] += colIt.val() * disp;
+                else
+                {
+                    addConstraintDisplacement_impl(d, i, systemLHVector_buf, positionIntegrationFactor, colIt.index(), colIt.val());
+                }
             }
         }
     }
@@ -613,18 +632,15 @@ void LinearSolverConstraintCorrection<DataTypes>::addConstraintDisplacement(doub
 template<class DataTypes>
 void LinearSolverConstraintCorrection<DataTypes>::setConstraintDForce(double *df, int begin, int end, bool update)
 {
-
-
-    const MatrixDeriv& constraints = mstate->read(core::ConstMatrixDerivId::constraintJacobian())->getValue();
-    const unsigned int derivDim = Deriv::size();
-
-
     last_force = begin;
 
     if (!update)
         return;
 
     _new_force = true;
+
+    constexpr const auto derivDim = Deriv::total_size;
+    const MatrixDeriv& constraints = mstate->read(core::ConstMatrixDerivId::constraintJacobian())->getValue();
 
     // TODO => optimisation !!!
     for (int i = begin; i <= end; i++)