Use directly operator[] instead of BaseMatrix::element() (and avoid using virtual functions) which is used extensively in addConstraintDisplacement() which is called a lot through ConstraintSolvers.
I dont like the way how to detect if the vector is FullVector at run-time (using dynamic_cast) but I did not see a fast way to get this info at compile-time.

Bench (3instruments_collis.scn from BeamAdapter), 5000 steps, MSVC2022:
- before: `74.5033 s ( 67.1111 FPS)`
- after: `61.728 s ( 81.0005 FPS)`

A nice 20% speedup




\+ some cleanups
 - color array was too long (messing IDE)
 - print only in dev-mode in LCPConstraintSolver if it did not reach convergence (bothersome)






______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).