@fredroy 2022-09-05
> 



> I just had a look on the implementation of `FullVector`. I notice the following differences between `operator[](Index i)` and `element(Index i)`:
> 
>     * `element(Index i)` is virtual
> 
>     * `element(Index i)` returns necessarily a `SReal` independent from the type of `FullVector`. It may lead to a type conversion therefore an impact on the performances. But usually, all the components are compiled with `SReal`, so I am not sure there is a conversion here. So it should not affect the performances.
> 
> 
> I am confused here, because I cannot believe that the speedup you get is only because of the virtual function. I guess you can try again your benchmark, but instead of using `operator[](Index i)`, use `FullVector<Real>::element(Index i)`. Note that I force the `element` function to be the one in `FullVector<Real>`.

with  `FullVector<Real>::element(Index i)`:
` 5000 iterations done in 59.448 s ( 84.1071 FPS)` more or less same as with the operator[] then