@fredroy 2022-09-06
FYI, in the bench, this function is called 7938654120 times. 😵